pyproject.toml
[project]
name = "llm-organic-company-history"
version = "0.1.0"
authors = [
    { name="Alex Coleman", email="coleman.g.alex@gmail.com" }
]
description = "Uses LLMs to generate an organic company history and HR data"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ollama>=0.3.3",
    "polars-lts-cpu>=1.12.0",
    "pandera[polars]>=0.20.4",
]

[project.optional-dependencies]
dev = [
    "ipython",
    "ipdb",
    "pdbpp",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

